+++ Rafy/Rafy/Domain/EntityRepository/Caching/CacheInstances.cs
@@ -8,6 +8,7 @@
  * 
  * 历史记录：
  * 创建文件 胡庆访 2011
+ * 修改文件 宋军瑞 20170112 -- 增加 PerHttpRequestCache 单例属性。
  * 
 *******************************************************/
